Add no-results behavior
If no audio files are detected on the current page, a prompt will appear explaining that no audio files are found and ask the user what they would ilke to hear. If user clicks "ok" then a tab will open with a custom google search to find music with the users input.